+++ backend/src/main/java/com/recinven/recinvenbackend/dto/ProductDto.java
@@ -2,5 +2,5 @@ package com.recinven.recinvenbackend.dto;
 
 public class ProductDto {
     public String description;
-    public float currentPrice;
+    public double currentPrice;
 }
